+++ ydb/library/workload/vector/registrar.cpp
@@ -1,4 +1,4 @@
-#include "vector.h"
+#include "vector_workload_params.h"
 #include <ydb/library/workload/abstract/workload_factory.h>
 
 namespace NYdbWorkload {
