--- ydb/library/workload/vector/vector.cpp
@@ -1,701 +1,0 @@
-#include "vector.h"
-
-#include <util/datetime/base.h>
-#include <util/generic/serialized_enum.h>
-
-#include <format>
-#include <string>
-
-#include <algorithm>
-
-namespace NYdbWorkload {
-
-// Initialize thread local atomic
-thread_local std::atomic<size_t> TVectorWorkloadGenerator::ThreadLocalTargetIndex{0};
-
-// Utility function to get metric info for SQL query
-// Returns a tuple of (function_name, is_ascending)
-std::tuple<std::string, bool> GetMetricInfo(NYdb::NTable::TVectorIndexSettings::EMetric metric) {
-    switch (metric) {
-        case NYdb::NTable::TVectorIndexSettings::EMetric::InnerProduct:
-            return {"InnerProductSimilarity", false}; // Similarity, higher is better (DESC)
-        
-        case NYdb::NTable::TVectorIndexSettings::EMetric::CosineSimilarity:
-            return {"CosineSimilarity", false}; // Similarity, higher is better (DESC)
-            
-        case NYdb::NTable::TVectorIndexSettings::EMetric::CosineDistance:
-            return {"CosineDistance", true}; // Distance, lower is better (ASC)
-            
-        case NYdb::NTable::TVectorIndexSettings::EMetric::Manhattan:
-            return {"ManhattanDistance", true}; // Distance, lower is better (ASC)
-            
-        case NYdb::NTable::TVectorIndexSettings::EMetric::Euclidean:
-            return {"EuclideanDistance", true}; // Distance, lower is better (ASC)
-            
-        case NYdb::NTable::TVectorIndexSettings::EMetric::Unspecified:
-        default:
-            Y_ABORT("Unspecified metric");
-    }
-}
-
-
-// Utility function to create select query
-std::string MakeSelect(const char* tableName, const char* indexName, const char* prefixColumn, 
-                     size_t kmeansTreeClusters, NYdb::NTable::TVectorIndexSettings::EMetric metric) {
-    auto [functionName, isAscending] = GetMetricInfo(metric);
-    
-    return std::format(R"_(--!syntax_v1
-        DECLARE $Embedding as String;
-        {0}
-        pragma ydb.KMeansTreeSearchTopSize="{1}";
-        SELECT id
-        FROM {2}
-        {3}
-        {4}
-        ORDER BY Knn::{5}(embedding, $Embedding) {6}
-        LIMIT $K;
-    )_", 
-        prefixColumn ? "DECLARE $PrefixValue as Int64;" : "",
-        kmeansTreeClusters,
-        tableName,
-        indexName ? std::format("VIEW {0}", indexName) : "",
-        prefixColumn ? std::format("WHERE {0} = $PrefixValue", prefixColumn) : "",
-        functionName,
-        isAscending ? "ASC" : "DESC"
-    );
-}
-
-
-// Utility function to create parameters for select query
-NYdb::TParams MakeSelectParams(const std::string& embeddingBytes, std::optional<i64> prefixValue, ui64 topK) {
-    NYdb::TParamsBuilder paramsBuilder;
-    
-    paramsBuilder.AddParam("$Embedding").String(embeddingBytes).Build();
-    paramsBuilder.AddParam("$K").Uint64(topK).Build();
-
-    if (prefixValue.has_value()) {
-        paramsBuilder.AddParam("$PrefixValue").Int64(*prefixValue).Build();
-    }
-
-    return paramsBuilder.Build();
-}
-
-// TVectorRecallEvaluator implementation
-TVectorRecallEvaluator::TVectorRecallEvaluator() {}
-
-TVectorRecallEvaluator::~TVectorRecallEvaluator() {
-    if (ProcessedTargets)
-        Cout << "Average recall: " << GetAverageRecall() << Endl;
-}
-
-void TVectorRecallEvaluator::SampleExistingVectors(const char* tableName, const char* prefixColumn, size_t targetCount, NYdb::NQuery::TQueryClient& queryClient) {
-    Cout << "Sampling " << targetCount << " vectors from dataset..." << Endl;
-    
-    // Create a local random generator
-    std::random_device rd;
-    std::mt19937_64 gen(rd());
-       
-    // First query to get min and max ID range
-    std::string rangeQuery = std::format(R"_(--!syntax_v1
-        SELECT Unwrap(MIN(id)) as min_id, Unwrap(MAX(id)) as max_id FROM {0};
-    )_", tableName);
-    
-    // Execute the range query
-    std::optional<NYdb::TResultSet> rangeResultSet;
-    NYdb::NStatusHelpers::ThrowOnError(queryClient.RetryQuerySync([&rangeQuery, &rangeResultSet](NYdb::NQuery::TSession session) {
-        auto result = session.ExecuteQuery(
-            rangeQuery,
-            NYdb::NQuery::TTxControl::NoTx())
-            .GetValueSync();
-        
-        Y_ABORT_UNLESS(result.IsSuccess(), "Range query failed: %s", result.GetIssues().ToString().c_str());
-        
-        rangeResultSet = result.GetResultSet(0);
-        return result;
-    }));
-    
-    // Parse the range result
-    NYdb::TResultSetParser rangeParser(*rangeResultSet);
-    Y_ABORT_UNLESS(rangeParser.TryNextRow());
-    ui64 minId = rangeParser.ColumnParser("min_id").GetUint64();
-    ui64 maxId = rangeParser.ColumnParser("max_id").GetUint64();
-    
-    Y_ABORT_UNLESS(minId <= maxId, "Invalid ID range in the dataset: min=%lu, max=%lu", minId, maxId);
-    
-    // Query to get total count of vectors in the table
-    std::string countQuery = std::format(R"_(--!syntax_v1
-        SELECT Unwrap(COUNT(*)) FROM {0};
-    )_", tableName);
-    
-    // Execute the count query
-    std::optional<NYdb::TResultSet> countResultSet;
-    NYdb::NStatusHelpers::ThrowOnError(queryClient.RetryQuerySync([&countQuery, &countResultSet](NYdb::NQuery::TSession session) {
-        auto result = session.ExecuteQuery(
-            countQuery,
-            NYdb::NQuery::TTxControl::NoTx())
-            .GetValueSync();
-        
-        Y_ABORT_UNLESS(result.IsSuccess(), "Count query failed: %s", result.GetIssues().ToString().c_str());
-        
-        countResultSet = result.GetResultSet(0);
-        return result;
-    }));
-    
-    // Parse the count result
-    NYdb::TResultSetParser countParser(*countResultSet);
-    Y_ABORT_UNLESS(countParser.TryNextRow());
-    ui64 totalVectors = countParser.ColumnParser(0).GetUint64();
-    
-    Y_ABORT_UNLESS(totalVectors > 0, "No vectors found in the dataset");
-    
-    // If we have fewer vectors than requested targets, adjust targetCount
-    if (totalVectors < targetCount) {
-        Cerr << "Warning: Requested " << targetCount << " targets but only " << totalVectors 
-             << " vectors exist in the dataset. Using all available vectors." << Endl;
-        targetCount = totalVectors;
-    }
-    
-    // Generate random IDs in the range
-    std::set<ui64> sampleIds;
-    std::uniform_int_distribution<ui64> idDist(minId, maxId);
-    size_t attemptsLimit = targetCount * 10; // Limit the number of attempts to avoid infinite loops
-    size_t attempts = 0;
-    
-    // Reserve space for targets
-    SelectTargets.clear();
-    SelectTargets.reserve(targetCount);
-    
-    // Keep sampling until we have enough targets or reach attempt limit
-    while (SelectTargets.size() < targetCount && attempts < attemptsLimit) {
-        ui64 randomId = idDist(gen);
-        
-        // Skip if we've already tried this ID
-        if (sampleIds.find(randomId) != sampleIds.end()) {
-            attempts++;
-            continue;
-        }
-        
-        sampleIds.insert(randomId);
-        attempts++;
-        
-        // Create query string based on whether we have a prefix column
-        std::string vectorQuery;
-        if (prefixColumn) {
-            vectorQuery = std::format(R"_(--!syntax_v1
-                SELECT id, Unwrap(embedding) as embedding, Unwrap({0}) as prefix_value 
-                FROM {1} 
-                WHERE id = {2};
-            )_", prefixColumn, tableName, randomId);
-        } else {
-            vectorQuery = std::format(R"_(--!syntax_v1
-                SELECT id, Unwrap(embedding) as embedding 
-                FROM {0} 
-                WHERE id = {1};
-            )_", tableName, randomId);
-        }
-        
-        std::optional<NYdb::TResultSet> vectorResultSet;
-        NYdb::NStatusHelpers::ThrowOnError(queryClient.RetryQuerySync([&vectorQuery, &vectorResultSet](NYdb::NQuery::TSession session) {
-            auto result = session.ExecuteQuery(
-                vectorQuery,
-                NYdb::NQuery::TTxControl::NoTx())
-                .GetValueSync();
-            
-            Y_ABORT_UNLESS(result.IsSuccess(), "Vector query failed: %s", result.GetIssues().ToString().c_str());
-            
-            vectorResultSet = result.GetResultSet(0);
-            return result;
-        }));
-        
-        // Parse the vector result
-        NYdb::TResultSetParser vectorParser(*vectorResultSet);
-        if (!vectorParser.TryNextRow()) {
-            // ID doesn't exist, try another random ID
-            continue;
-        }
-        
-        ui64 id = vectorParser.ColumnParser("id").GetUint64();
-        std::string embeddingBytes = vectorParser.ColumnParser("embedding").GetString();
-        
-        // Ensure we got a valid embedding
-        Y_ABORT_UNLESS(!embeddingBytes.empty(), "Empty embedding retrieved for id %" PRIu64, id);
-        
-        // Create target with the sampled vector
-        TSelectTarget target;
-        target.EmbeddingBytes = std::move(embeddingBytes);
-        target.Etalons.push_back(id);  // The vector's own ID is an etalon
-        
-        // Get prefix value if column was specified
-        if (prefixColumn) {
-            const auto& prefixCell = vectorParser.ColumnParser("prefix_value");
-            NYdb::EPrimitiveType primitiveType = prefixCell.GetPrimitiveType();
-            
-            // Handle different integer types based on primitive type
-            switch (primitiveType) {
-                case NYdb::EPrimitiveType::Int8:
-                    target.PrefixValue = prefixCell.GetInt8();
-                    break;
-                case NYdb::EPrimitiveType::Int16:
-                    target.PrefixValue = prefixCell.GetInt16();
-                    break;
-                case NYdb::EPrimitiveType::Int32:
-                    target.PrefixValue = prefixCell.GetInt32();
-                    break;
-                case NYdb::EPrimitiveType::Int64:
-                    target.PrefixValue = prefixCell.GetInt64();
-                    break;
-                case NYdb::EPrimitiveType::Uint8:
-                    target.PrefixValue = prefixCell.GetUint8();
-                    break;
-                case NYdb::EPrimitiveType::Uint16:
-                    target.PrefixValue = prefixCell.GetUint16();
-                    break;
-                case NYdb::EPrimitiveType::Uint32:
-                    target.PrefixValue = prefixCell.GetUint32();
-                    break;
-                case NYdb::EPrimitiveType::Uint64: {
-                    ui64 uvalue = prefixCell.GetUint64();
-                    Y_ABORT_UNLESS(uvalue <= static_cast<ui64>(std::numeric_limits<i64>::max()), 
-                                  "Prefix value %" PRIu64 " is too large for i64", uvalue);
-                    target.PrefixValue = static_cast<i64>(uvalue);
-                    break;
-                }
-                default:
-                    Y_ABORT_UNLESS(false, "Unexpected primitive type for prefix column: %d", 
-                                  static_cast<int>(primitiveType));
-            }
-        }
-        
-        SelectTargets.push_back(std::move(target));
-    }
-    
-    if (SelectTargets.size() < targetCount) {
-        Cerr << "Warning: Could only sample " << SelectTargets.size() << " vectors after " << attempts << " attempts." << Endl;
-    } else {
-        Cout << "Successfully sampled " << SelectTargets.size() << " vectors from the dataset." << Endl;
-    }
-    Y_ABORT_UNLESS(!SelectTargets.empty(), "Failed to sample any vectors from the dataset");
-}
-
-void TVectorRecallEvaluator::FillEtalons(const char* tableName, const char* prefixColumn, NYdb::NTable::TVectorIndexSettings::EMetric metric, size_t topK, NYdb::NQuery::TQueryClient& queryClient) {
-    Cout << "Recall initialization..." << Endl;
-    
-    // Prepare the query template
-    std::string queryTemplate = MakeSelect(tableName, nullptr, prefixColumn, 0, metric);
-    
-    // Maximum number of concurrent queries to execute
-    const size_t MAX_CONCURRENT_QUERIES = 20;
-    
-    // Process targets in batches
-    for (size_t batchStart = 0; batchStart < SelectTargets.size(); batchStart += MAX_CONCURRENT_QUERIES) {
-        size_t batchEnd = std::min(batchStart + MAX_CONCURRENT_QUERIES, SelectTargets.size());
-        
-        // Start async queries for this batch
-        std::vector<std::pair<size_t, NYdb::NQuery::TAsyncExecuteQueryResult>> asyncQueries;
-        asyncQueries.reserve(batchEnd - batchStart);
-        
-        for (size_t i = batchStart; i < batchEnd; i++) {
-            const auto& targetEmbedding = GetTargetEmbedding(i);
-            std::optional<i64> prefixValue;
-            if (prefixColumn) {
-                prefixValue = GetPrefixValue(i);
-            };
-
-            NYdb::TParams params = MakeSelectParams(targetEmbedding, prefixValue, topK);
-            
-            auto asyncResult = queryClient.RetryQuery([queryTemplate, params](NYdb::NQuery::TSession session) {
-                return session.ExecuteQuery(
-                    queryTemplate,
-                    NYdb::NQuery::TTxControl::NoTx(),
-                    params);
-            });
-            
-            asyncQueries.emplace_back(i, std::move(asyncResult));
-        }
-        
-        // Wait for all queries in this batch to complete
-        for (auto& [targetIndex, asyncResult] : asyncQueries) {
-            auto result = asyncResult.GetValueSync();
-            Y_ABORT_UNLESS(result.IsSuccess(), "Query failed for target %zu: %s", 
-                          targetIndex, result.GetIssues().ToString().c_str());
-            
-            auto resultSet = result.GetResultSet(0);
-            NYdb::TResultSetParser parser(resultSet);
-            
-            // Clear and prepare etalons for this target
-            SelectTargets[targetIndex].Etalons.clear();
-            SelectTargets[targetIndex].Etalons.reserve(topK);
-            
-            // Extract all IDs from the result set
-            while (parser.TryNextRow()) {
-                ui64 id = parser.ColumnParser("id").GetUint64();
-                SelectTargets[targetIndex].Etalons.push_back(id);
-            }
-            if (SelectTargets[targetIndex].Etalons.empty()) {
-                Cerr << "Warning: target " << targetIndex << " have empty etalon sets" << Endl;
-            }
-        }
-        
-        // Log progress for large datasets
-        if (SelectTargets.size() > 100 && (batchEnd % 100 == 0 || batchEnd == SelectTargets.size())) {
-            Cout << "Processed " << batchEnd << " of " << SelectTargets.size() << " targets..." << Endl;
-        }
-    }
-    
-    Cout << "Recall initialization completed for " << SelectTargets.size() << " targets." << Endl;
-}
-
-const std::string& TVectorRecallEvaluator::GetTargetEmbedding(size_t index) const {
-    return SelectTargets.at(index).EmbeddingBytes;
-}
-
-const std::vector<ui64>& TVectorRecallEvaluator::GetTargetEtalons(size_t index) const {
-    return SelectTargets.at(index).Etalons;
-}
-
-void TVectorRecallEvaluator::AddRecall(double recall) {
-    std::lock_guard<std::mutex> lock(Mutex);
-    TotalRecall += recall;
-    ProcessedTargets++;
-}
-
-double TVectorRecallEvaluator::GetAverageRecall() const {
-    return ProcessedTargets > 0 ? TotalRecall / ProcessedTargets : 0.0;
-}
-
-double TVectorRecallEvaluator::GetTotalRecall() const {
-    return TotalRecall;
-}
-
-size_t TVectorRecallEvaluator::GetProcessedTargets() const {
-    return ProcessedTargets;
-}
-
-size_t TVectorRecallEvaluator::GetTargetCount() const {
-    return SelectTargets.size();
-}
-
-i64 TVectorRecallEvaluator::GetPrefixValue(size_t targetIndex) const {
-    return SelectTargets.at(targetIndex).PrefixValue;
-}
-
-// Process query results
-void TVectorRecallEvaluator::ProcessQueryResult(const NYdb::NQuery::TExecuteQueryResult& queryResult, size_t targetIndex, bool verbose) {
-    if (!queryResult.IsSuccess()) {
-        // Ignore the error. It's printed in the verbose mode
-        return;
-    }
-    
-    // Get the result set
-    auto resultSet = queryResult.GetResultSet(0);
-    NYdb::TResultSetParser parser(resultSet);
-    
-    // Extract IDs from index search results
-    std::vector<ui64> indexResults;
-    while (parser.TryNextRow()) {
-        ui64 id = parser.ColumnParser("id").GetUint64();
-        indexResults.push_back(id);
-    }
-    
-    // Get the etalons for the specified target
-    const auto& etalons = GetTargetEtalons(targetIndex);
-    std::unordered_set<ui64> etalonSet(etalons.begin(), etalons.end());
-
-    // Check if target ID is first in results
-    if (!indexResults.empty() && !etalons.empty()) {
-        ui64 targetId = etalons[0]; // First etalon is the target ID itself
-        
-        if (verbose && indexResults[0] != targetId) {
-            Cerr << "Warning: Target ID " << targetId << " is not the first result for target " 
-                 << targetIndex << ". Found " << indexResults[0] << " instead." << Endl;
-        }
-        
-        // Calculate recall
-        size_t relevantRetrieved = 0;
-        for (const auto& id : indexResults) {
-            if (etalonSet.count(id)) {
-                relevantRetrieved++;
-            }
-        }
-        
-        // Calculate recall for this target
-        double recall = etalons.empty() ? 0.0 : static_cast<double>(relevantRetrieved) / etalons.size();
-        AddRecall(recall);
-
-        // Add warning when zero relevant results found
-        if (verbose && relevantRetrieved == 0 && !indexResults.empty()) {
-            Cerr << "Warning: Zero relevant results for target " << targetIndex << Endl;
-        }
-    } else {
-        // Handle empty results or empty etalons
-        if (verbose)
-            Cerr << "Warning: Empty results or etalons for target " << targetIndex << Endl;
-    }
-}
-
-// TVectorWorkloadGenerator implementation
-TVectorWorkloadGenerator::TVectorWorkloadGenerator(const TVectorWorkloadParams* params)
-    : TBase(params)
-{
-}
-
-std::string TVectorWorkloadGenerator::GetDDLQueries() const {
-    return std::format(R"_(--!syntax_v1
-        CREATE TABLE `{0}/{1}`(
-            id Uint64,
-            embedding Utf8,
-            PRIMARY KEY(id)) 
-        WITH (
-            AUTO_PARTITIONING_BY_SIZE = ENABLED,
-            AUTO_PARTITIONING_PARTITION_SIZE_MB = 500,
-            AUTO_PARTITIONING_BY_LOAD = ENABLED
-        )
-    )_", Params.DbPath.c_str(), Params.TableName.c_str());
-}
-
-TQueryInfoList TVectorWorkloadGenerator::GetInitialData() {
-    return {};
-}
-
-TVector<std::string> TVectorWorkloadGenerator::GetCleanPaths() const {
-    return {"vector"};
-}
-
-TQueryInfoList TVectorWorkloadGenerator::GetWorkload(int type) {
-    switch (static_cast<EType>(type)) {
-        case EType::Upsert:
-            return Upsert();
-        case EType::Select:
-            return Select();
-        default:
-            return TQueryInfoList();
-    }
-}
-
-TVector<IWorkloadQueryGenerator::TWorkloadType> TVectorWorkloadGenerator::GetSupportedWorkloadTypes() const {
-    TVector<TWorkloadType> result;
-    result.emplace_back(static_cast<int>(EType::Upsert), "upsert", "Upsert vector rows in the table");
-    result.emplace_back(static_cast<int>(EType::Select), "select", "Retrieve top-K vectors");
-    return result;
-}
-
-TQueryInfoList TVectorWorkloadGenerator::Upsert() {
-    // Not implemented yet
-    return {};
-}
-
-size_t TVectorWorkloadGenerator::GetNextTargetIndex(size_t currentIndex) const {
-    return (currentIndex + 1) % Params.VectorRecallEvaluator->GetTargetCount();
-}
-
-TQueryInfoList TVectorWorkloadGenerator::Select() {
-    // Initialize thread-local target index if needed
-    if (ThreadLocalTargetIndex.load() == 0) {
-        std::random_device rd;
-        std::mt19937 gen(rd());
-        std::uniform_int_distribution<size_t> dist(0, Params.VectorRecallEvaluator->GetTargetCount() - 1);
-        ThreadLocalTargetIndex.store(dist(gen));
-    }
-    
-    // Get current target index
-    size_t targetIndex = ThreadLocalTargetIndex.load();
-
-    // Create the query string
-    std::string query = MakeSelect(
-        Params.TableName.c_str(), 
-        Params.IndexName.c_str(), 
-        Params.PrefixColumn.has_value() ? Params.PrefixColumn->c_str() : nullptr,
-        Params.KmeansTreeSearchClusters,
-        Params.Metric
-    );
-    
-    // Get the embedding for the specified target
-    const auto& targetEmbedding = Params.VectorRecallEvaluator->GetTargetEmbedding(targetIndex);
-    
-    // Get the prefix value if needed
-    std::optional<i64> prefixValue;
-    if (Params.PrefixColumn.has_value()) {
-        prefixValue = Params.VectorRecallEvaluator->GetPrefixValue(targetIndex);
-    }
-    
-    NYdb::TParams params = MakeSelectParams(targetEmbedding, prefixValue, Params.TopK);
-    
-    // Create the query info with a callback that captures the target index
-    TQueryInfo queryInfo(query, std::move(params));
-    if (Params.Recall) {
-        queryInfo.GenericQueryResultCallback = std::bind(
-            &TVectorWorkloadGenerator::RecallCallback, 
-            this, 
-            std::placeholders::_1, 
-            targetIndex
-        );
-    }
-    
-    // Update the thread-local index for the next query
-    ThreadLocalTargetIndex.store(GetNextTargetIndex(targetIndex));
-    
-    return TQueryInfoList(1, queryInfo);
-}
-
-
-void TVectorWorkloadGenerator::RecallCallback(NYdb::NQuery::TExecuteQueryResult queryResult, size_t targetIndex) {
-    // Use the evaluator to process the result and calculate recall
-    Params.VectorRecallEvaluator->ProcessQueryResult(queryResult, targetIndex, Params.Verbose);
-}
-
-// TVectorWorkloadParams implementation
-void TVectorWorkloadParams::ConfigureOpts(NLastGetopt::TOpts& opts, const ECommandType commandType, int workloadType) {
-    auto addCommonParam = [&]() {
-        opts.AddLongOption( "table", "Table name.")
-            .DefaultValue("vector_index_workload").StoreResult(&TableName);
-        opts.AddLongOption( "index", "Index name.")
-            .DefaultValue("index").StoreResult(&IndexName);
-    };
-
-    auto addInitParam = [&]() {
-        opts.AddLongOption( "rows", "Number of vectors to init the table.")
-            .Required().StoreResult(&VectorInitCount);
-        opts.AddLongOption( "distance", "Distance/similarity function")
-            .Required().StoreResult(&Distance);
-        opts.AddLongOption( "vector-type", "Type of vectors")
-            .Required().StoreResult(&VectorType);
-        opts.AddLongOption( "vector-dimension", "Vector dimension.")
-            .Required().StoreResult(&VectorDimension);
-        opts.AddLongOption( "kmeans-tree-levels", "Number of levels in the kmeans tree")
-            .Required().StoreResult(&KmeansTreeLevels);
-        opts.AddLongOption( "kmeans-tree-clusters", "Number of cluster in kmeans")
-            .Required().StoreResult(&KmeansTreeClusters);
-
-    };
-
-    auto addUpsertParam = [&]() {
-    };
-
-    auto addSelectParam = [&]() {
-        opts.AddLongOption( "targets", "Number of vectors to search as targets.")
-            .DefaultValue(100).StoreResult(&Targets);
-        opts.AddLongOption( "top-k", "Number of top vector to return.")
-            .DefaultValue(5).StoreResult(&TopK);
-        opts.AddLongOption( "kmeans-tree-clusters", "Number of top clusters during search.")
-            .DefaultValue(1).StoreResult(&KmeansTreeSearchClusters);
-        opts.AddLongOption( "recall", "Measure recall metrics. It trains on 'targets' vector by bruce-force search.")
-            .StoreTrue(&Recall); 
-    };
-
-    switch (commandType) {
-    case TWorkloadParams::ECommandType::Init:
-        addCommonParam();
-        addInitParam();
-        break;
-    case TWorkloadParams::ECommandType::Run:
-        addCommonParam();
-        switch (static_cast<TVectorWorkloadGenerator::EType>(workloadType)) {
-        case TVectorWorkloadGenerator::EType::Upsert:
-            addUpsertParam();
-            break;
-        case TVectorWorkloadGenerator::EType::Select:
-            addSelectParam();
-            break;
-        }
-        break;
-    default:
-        break;
-    }
-}
-
-void TVectorWorkloadParams::GetIndexInfo() {
-    // Build the full path to the table
-    std::string tablePath = DbPath + "/" + TableName;
-
-    auto session = TableClient->GetSession().ExtractValueSync().GetSession();
-    auto describeTableResult = session.DescribeTable(tablePath).GetValueSync();
-    Y_ABORT_UNLESS(describeTableResult.IsSuccess(), "DescribeTable failed: %s", describeTableResult.GetIssues().ToString().c_str());
-    
-    // Get the table description
-    const auto& tableDescription = describeTableResult.GetTableDescription();
-    
-    // Find the specified index
-    bool indexFound = false;
-    
-    for (const auto& index : tableDescription.GetIndexDescriptions()) {
-        if (index.GetIndexName() == IndexName) {
-            indexFound = true;
-            
-            // Check if we have more than one column (indicating a prefixed index)
-            const auto& keyColumns = index.GetIndexColumns();
-            if (keyColumns.size() > 1) {
-                // The first column is the prefix column, the last column is the embedding
-                PrefixColumn = keyColumns[0];
-            }
-            
-            // Extract the distance metric from index settings
-            const auto& indexSettings = std::get<NYdb::NTable::TKMeansTreeSettings>(index.GetIndexSettings());
-            Metric = indexSettings.Settings.Metric;
-            VectorDimension = indexSettings.Settings.VectorDimension;
-            
-            break;
-        }
-    }
-
-    // If a prefix column was found, verify that it has an integer type
-    if (PrefixColumn.has_value()) {
-        // Find the column in table schema
-        for (const auto& column : tableDescription.GetColumns()) {
-            if (column.Name == PrefixColumn.value()) {
-                // Check column type
-                const NYdb::TType& typeInfo = column.Type;
-                
-                // Check if it's an integer type
-                Y_ABORT_UNLESS(typeInfo.ToString().contains("int") || typeInfo.ToString().contains("Int"), 
-                    "Prefix column '%s' in index '%s' must be an integer type. Found type: %s", 
-                    PrefixColumn->c_str(), IndexName.c_str(), typeInfo.ToString().c_str());          
-            }
-        }
-    }
-    
-    Y_ABORT_UNLESS(indexFound, "Index %s not found in table %s", IndexName.c_str(), TableName.c_str());
-}
-
-void TVectorWorkloadParams::Validate(const ECommandType commandType, int workloadType) {
-    switch (commandType) {
-        case TWorkloadParams::ECommandType::Init:
-            break;
-        case TWorkloadParams::ECommandType::Run:
-            switch (static_cast<TVectorWorkloadGenerator::EType>(workloadType)) {
-                case TVectorWorkloadGenerator::EType::Upsert:
-                    break;
-                case TVectorWorkloadGenerator::EType::Select:
-                    GetIndexInfo();
-                    VectorRecallEvaluator = MakeHolder<TVectorRecallEvaluator>();
-                    const char* prefixColumn = PrefixColumn.has_value() ? PrefixColumn->c_str() : nullptr;
-                    VectorRecallEvaluator->SampleExistingVectors(TableName.c_str(), prefixColumn, Targets, *QueryClient);
-                    if (Recall) {
-                        VectorRecallEvaluator->FillEtalons(
-                            TableName.c_str(), 
-                            prefixColumn,
-                            Metric,
-                            TopK, 
-                            *QueryClient
-                        );
-                    }
-                    break;
-            }
-            break;
-        case TWorkloadParams::ECommandType::Clean:
-            break;
-        case TWorkloadParams::ECommandType::Root:
-            break;
-        case TWorkloadParams::ECommandType::Import:
-            break;
-    }
-}
-
-THolder<IWorkloadQueryGenerator> TVectorWorkloadParams::CreateGenerator() const {
-    return MakeHolder<TVectorWorkloadGenerator>(this);
-}
-
-TString TVectorWorkloadParams::GetWorkloadName() const {
-    return "vector";
-}
-
-} // namespace NYdbWorkload
